pyproject.toml
[project]
name = "room_project"
version = "0.1.0"
description = "A project to manage room reservations"
authors = [
    { name = "Matthieu GRIMM--KEMPF", email = "matthieu.grimm-kempf@uha.fr" },
    { name = "Lucas DAGON", email = "lucas.dagon@uha.fr" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSD-3-Clause" }
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "babel==2.17.0",
    "pillow==11.2.1",
    "tkcalendar==1.6.1",
    "typer"
]

[project.optional-dependencies]
test = ["pytest >= 8.1.1",
	"pytest-cov",
	"pytest-html"
]

[build-system]
requires = ["setuptools", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project.scripts]
room_project = "room_project.cli:application"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

